+++ plugins/kiali/src/components/VirtualList/VirtualList.tsx
@@ -1,6 +1,8 @@
 import * as React from 'react';
 
 import {
+  Box,
+  CircularProgress,
   Paper,
   SortDirection,
   Table,
@@ -16,6 +18,7 @@ import { kialiStyle } from '../../styles/StyleUtils';
 import { Namespace } from '../../types/Namespace';
 import { NamespaceInfo } from '../../types/NamespaceInfo';
 import { SortField } from '../../types/SortFilters';
+import { ENTITY } from '../../types/types';
 import { StatefulFiltersProps } from '../Filters/StatefulFilters';
 import { config, RenderResource, Resource, ResourceType } from './Config';
 import { VirtualItem } from './VirtualItem';
@@ -44,6 +47,7 @@ type VirtualListProps<R> = {
   tableToolbar?: React.ReactNode;
   type: string;
   view?: string;
+  loading: boolean;
 };
 
 export const VirtualList = <R extends RenderResource>(
@@ -70,6 +74,16 @@ export const VirtualList = <R extends RenderResource>(
   const { rows } = listProps;
   const typeDisplay =
     listProps.type === 'istio' ? 'Istio config' : listProps.type;
+
+  const tableEntityHeaderStyle: any = {
+    minWidth: '100px',
+    fontWeight: '700',
+    color: 'grey',
+    borderTop: '1px solid #d5d5d5',
+    borderBottom: '1px solid #d5d5d5',
+    whiteSpace: 'nowrap',
+    padding: '15px',
+  };
 
   const tableHeaderStyle: any = {
     minWidth: '120px',
@@ -141,7 +155,11 @@ export const VirtualList = <R extends RenderResource>(
                   <TableCell
                     key={`column_${index}`}
                     align="center"
-                    style={tableHeaderStyle}
+                    style={
+                      listProps.view === ENTITY
+                        ? tableEntityHeaderStyle
+                        : tableHeaderStyle
+                    }
                     sortDirection={
                       column.sortable && orderBy === column.title.toLowerCase()
                         ? order
@@ -158,14 +176,28 @@ export const VirtualList = <R extends RenderResource>(
                         handleRequestSort(e, column.title.toLowerCase())
                       }
                     >
-                      {column.title.toUpperCase()}
+                      {listProps.view === ENTITY &&
+                      column.title === 'Configuration'
+                        ? 'CONFIG'
+                        : column.title.toUpperCase()}
                     </TableSortLabel>
                   </TableCell>
                 ))}
               </TableRow>
             </TableHead>
             <TableBody>
-              {listProps.rows.length > 0 ? (
+              {/* eslint-disable-next-line no-nested-ternary */}
+              {listProps.loading === true ? (
+                <Box
+                  display="flex"
+                  justifyContent="center"
+                  alignItems="center"
+                  minHeight="10vh"
+                  marginLeft="60vh"
+                >
+                  <CircularProgress />
+                </Box>
+              ) : listProps.rows.length > 0 ? (
                 stableSort(rows, getComparator()).map(
                   (row: RenderResource, index: number) => (
                     <VirtualItem