+++ plugins/orchestrator-backend/src/service/api/v1.ts
@@ -3,6 +3,7 @@ import express from 'express';
 import {
   AssessedProcessInstance,
   ProcessInstance,
+  ProcessInstanceVariables,
   WorkflowDefinition,
   WorkflowExecutionResponse,
   WorkflowOverview,
@@ -108,7 +109,7 @@ export class V1 {
   }
 
   public async executeWorkflow(
-    reqBody: Record<string, any>,
+    inputData: ProcessInstanceVariables,
     definitionId: string,
     businessKey: string | undefined,
   ): Promise<WorkflowExecutionResponse> {
@@ -124,7 +125,7 @@ export class V1 {
     }
     const executionResponse = await this.orchestratorService.executeWorkflow({
       definitionId: definitionId,
-      inputData: reqBody,
+      inputData,
       serviceUrl: definition.serviceUrl,
       businessKey,
       cacheHandler: 'throw',
@@ -153,6 +154,53 @@ export class V1 {
       instanceId,
       cacheHandler: 'throw',
     });
+  }
+
+  public async retriggerInstanceInError(
+    instanceId: string,
+    inputData: ProcessInstanceVariables,
+  ): Promise<WorkflowExecutionResponse> {
+    const instance = await this.orchestratorService.fetchInstance({
+      instanceId,
+      cacheHandler: 'throw',
+    });
+
+    if (!instance?.serviceUrl) {
+      throw new Error(`Couldn't fetch process instance ${instanceId}`);
+    }
+
+    if (instance.state !== 'ERROR') {
+      throw new Error(
+        `Can't retrigger an instance on ${instance.state} state.`,
+      );
+    }
+
+    const isUpdateInstanceInputDataOk =
+      await this.orchestratorService.updateInstanceInputData({
+        definitionId: instance.processId,
+        instanceId,
+        inputData,
+        serviceUrl: instance.serviceUrl,
+        cacheHandler: 'throw',
+      });
+
+    if (!isUpdateInstanceInputDataOk) {
+      throw new Error(`Couldn't update instance input data for ${instanceId}`);
+    }
+
+    const isRetriggerInstanceInErrorOk =
+      await this.orchestratorService.retriggerInstanceInError({
+        definitionId: instance.processId,
+        instanceId,
+        serviceUrl: instance.serviceUrl,
+        cacheHandler: 'throw',
+      });
+
+    if (!isRetriggerInstanceInErrorOk) {
+      throw new Error(`Couldn't retrigger instance in error for ${instanceId}`);
+    }
+
+    return { id: instanceId };
   }
 
   public extractQueryParam(