+++ plugins/rbac-backend/src/service/enforcer-delegate.test.ts
@@ -173,6 +173,7 @@ describe('EnforcerDelegate', () => {
       logger,
       tokenManagerMock,
       catalogDBClient,
+      config,
     );
     enf.setRoleManager(rm);
     enf.enableAutoBuildRoleLinks(false);