+++ plugins/kiali-backend/CHANGELOG.md
@@ -1,3 +1,11 @@
+## @janus-idp/backstage-plugin-kiali-backend [1.10.7](https://github.com/janus-idp/backstage-plugins/compare/@janus-idp/backstage-plugin-kiali-backend@1.10.6...@janus-idp/backstage-plugin-kiali-backend@1.10.7) (2024-04-25)
+
+
+
+### Dependencies
+
+* **@janus-idp/cli:** upgraded to 1.8.1
+
 ## @janus-idp/backstage-plugin-kiali-backend [1.10.6](https://github.com/janus-idp/backstage-plugins/compare/@janus-idp/backstage-plugin-kiali-backend@1.10.5...@janus-idp/backstage-plugin-kiali-backend@1.10.6) (2024-04-15)